+++ commands/sendgrid.js
@@ -1,6 +1,6 @@
 import * as store from "../lib/store.js"
 import { isTriggerfishOfficeIp } from "../lib/triggerfish.js"
-import { writeInfo, writeStep, writeSuccess } from "../lib/write.js"
+import { writeError, writeInfo, writeStep, writeSuccess } from "../lib/write.js"
 import setupSendgridAccount from "../tasks/services/sendgrid.js"
 
 export async function createSendGrid() {
@@ -11,9 +11,13 @@ export async function createSendGrid() {
 
   let sendgridApiKey = store.get("sendgridApiKey")
 
-  writeSuccess("SendGrid account created!")
-  writeInfo(`Add the following values to your vault file or .env:
-    tf_smtp_username: "apikey",
-    tf_smtp_password: ${sendgridApiKey}
-`)
+  if (sendgridApiKey) {
+    writeSuccess("SendGrid account created!")
+    writeInfo(`Add the following values to your vault file or .env:
+      tf_smtp_username: "apikey",
+      tf_smtp_password: ${sendgridApiKey}
+  `)
+  } else {
+    writeError("No Sendgrid account was created.")
+  }
 }