+++ index.js
@@ -7,6 +7,7 @@ import cloneLisaProject from "./commands/clone.js"
 import configure from "./commands/configure.js"
 import dbImport from "./commands/db.js"
 import init from "./commands/init.js"
+import { kinsta } from "./commands/kinsta.js"
 import { createPageComponent } from "./commands/pageComponent.js"
 import setupPath from "./commands/path.js"
 import { createSendGrid } from "./commands/sendgrid.js"
@@ -17,9 +18,10 @@ import exec from "./lib/exec.js"
 import { getSitesPath } from "./lib/path.js"
 import { set } from "./lib/store.js"
 import { checkLisaVersion } from "./lib/versions.js"
-import { kinsta } from "./commands/kinsta.js"
+import { wpUpdate } from "./commands/wordpressComposerUpdate.js"
 
 export const program = new Command()
+export const LISA_VERSION = "2.12.0"
 
 resetConf()
 checkNodeVersion()
@@ -34,6 +36,8 @@ async function initProgram() {
   set("initialPath", initialPath.stdout.trim())
 
   process.chdir(await getSitesPath())
+
+  program.version(LISA_VERSION, "-v, --version", "Output the current version")
 
   program
     .command("path")
@@ -110,6 +114,15 @@ async function initProgram() {
     .command("sendgrid create")
     .description("Create SendGrid account")
     .action(createSendGrid)
+
+  program
+    .command("wp update")
+    .description(
+      `Update WordPress and Composer dependencies.
+Make sure your standing in the folder where your composer.json file is located.
+    `
+    )
+    .action(wpUpdate)
 
   program.parse()
 }